Refactor code more to have PillGroup as a separate component.
