+++ assets/src/edit-story/components/library/panes/media/media3p/media3pPane.js
@@ -43,7 +43,7 @@ import useLibrary from '../../../useLibrary';
 import Flags from '../../../../../flags';
 import { PROVIDERS } from '../../../../../app/media/media3p/providerConfiguration';
 import resourceList from '../../../../../utils/resourceList';
-import Media3pCategories from './media3pCategories';
+import { PillGroup } from '../../shared';
 import paneId from './paneId';
 import ProviderTab from './providerTab';
 
@@ -179,7 +179,7 @@ function Media3pPane(props) {
         id={`provider-bottom-wrapper-${providerType}`}
       >
         {PROVIDERS[providerType].supportsCategories && (
-          <Media3pCategories
+          <PillGroup
             categories={state.categories.categories}
             selectedCategoryId={state.categories.selectedCategoryId}
             selectCategory={actions.selectCategory}
