+++ assets/src/edit-story/components/library/panes/shared/pillGroup/index.js
@@ -13,4 +13,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export { default } from './pill';
+export { default } from './pillGroup';
